// src/modules/notification/enums/notification-event.enum.ts
export enum NotificationEvent {
    REGISTRATION_OTP = 'REGISTRATION_OTP',
    RESEND_OTP = 'RESEND_OTP',
    FORGOT_PASSWORD_OTP = 'FORGOT_PASSWORD_OTP',
    PASSWORD_CHANGED = 'PASSWORD_CHANGED',
    PASSWORD_RESET = 'PASSWORD_RESET',
    NEW_LOGIN = 'NEW_LOGIN',
    SELLER_APPLICATION_SUBMITTED = 'SELLER_APPLICATION_SUBMITTED',
    SELLER_APPROVED = 'SELLER_APPROVED',
    SELLER_REJECTED = 'SELLER_REJECTED',
}